requirements.txt
alabaster==0.7.13
arrow==1.2.3
astroid==2.15.0
asttokens==2.2.1
atomicwrites==1.4.1
attrs==22.2.0
autopep8==1.6.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.0
binaryornot==0.4.4
black==23.1.0
bleach==6.0.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cookiecutter==2.1.1
cryptography==39.0.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
distlib==0.3.6
docstring-to-markdown==0.12
docutils==0.19
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.10.3
flake8==6.0.0
fonttools==4.39.2
idna==3.4
imagesize==1.4.1
importlib-metadata==6.1.0
importlib-resources==5.12.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
isort==5.12.0
jaraco.classes==3.2.3
jedi==0.18.2
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyterlab-pygments==0.2.2
keyring==23.13.1
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
more-itertools==9.1.0
mpmath==1.3.0
mypy-extensions==1.0.0
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
numpy==1.24.2
numpydoc==1.5.0
opencv-contrib-python==4.9.0.80
openpyxl==3.1.4
packaging==23.0
pandas==2.2.0
pandocfilters==1.5.0
paramiko==3.1.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.2.3
pyflakes==3.0.1
Pygments==2.14.0
pylint==2.17.0
pylint-venv==3.0.1
pyls-spyder==0.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
python-slugify==8.0.1
pytoolconfig==1.2.5
pytz==2024.1
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==25.0.2
QDarkStyle==3.0.3
qstylizer==0.2.2
QtAwesome==1.2.3
qtconsole==5.4.1
QtPy==2.3.0
requests==2.28.2
rope==1.7.0
Rtree==1.0.1
scikit-spatial==6.8.1
scipy==1.10.1
seaborn==0.13.2
SimpleITK==2.2.1
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.2
spyder-kernels==2.4.2
stack-data==0.6.2
sympy==1.11.1
text-unidecode==1.3
textdistance==4.5.0
three-merge==0.1.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2024.1
ujson==5.7.0
urllib3==1.26.15
virtualenv==20.21.0
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
whatthepatch==1.0.4
wrapt==1.15.0
yapf==0.32.0
zipp==3.15.0
